backend/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --allow-unsafe requirements/requirements.in
#
asgiref==3.7.2
    # via django
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via requests
cryptography==41.0.1
    # via social-auth-core
defusedxml==0.7.1
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.2
    # via
    #   -r requirements/requirements.in
    #   django-filter
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-extra-fields
    #   social-auth-app-django
django-filter==23.2
    # via -r requirements/requirements.in
django-templated-mail==1.1.1
    # via djoser
djangorestframework==3.14.0
    # via
    #   -r requirements/requirements.in
    #   djangorestframework-simplejwt
    #   drf-extra-fields
djangorestframework-simplejwt==5.2.2
    # via djoser
djoser==2.2.0
    # via -r requirements/requirements.in
drf-extra-fields==3.5.0
    # via -r requirements/requirements.in
filetype==1.2.0
    # via drf-extra-fields
gunicorn==20.1.0
    # via -r requirements/requirements.in
idna==3.4
    # via requests
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
pillow==9.5.0
    # via -r requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pycparser==2.21
    # via cffi
pyjwt==2.7.0
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
python-dotenv==1.0.0
    # via -r requirements/requirements.in
python3-openid==3.2.0
    # via social-auth-core
pytz==2023.3
    # via djangorestframework
requests==2.31.0
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==1.3.1
    # via social-auth-core
social-auth-app-django==5.2.0
    # via djoser
social-auth-core==4.4.2
    # via social-auth-app-django
sqlparse==0.4.4
    # via django
typing-extensions==4.7.1
    # via asgiref
urllib3==2.0.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
setuptools==68.0.0
    # via gunicorn
